// src/app/models/erc.ts
export interface ERC {
    ID_CUENTA_ERC?: string;
    CAMPO_1?: string;
    CAMPO_2?: string;
    CAMPO_3?: string;
    CAMPO_4?: string;
    CAMPO_5?: string;
    CAMPO_6?: string;
    CAMPO_7?: string;
    CAMPO_8?: string;
    CAMPO_9?: string;
    CAMPO_10?: string;
    CAMPO_11?: string;
    CAMPO_12?: string;
    CAMPO_13?: string;
    CAMPO_14?: string;
    CAMPO_15?: string;
    CAMPO_16?: string;
    CAMPO_17?: string;
    CAMPO_18?: string;
    CAMPO_19?: string;
    CAMPO_19_1?: string;
    CAMPO_20?: string;
    CAMPO_21?: string;
    CAMPO_21_1?: string;
    CAMPO_22?: string;
    CAMPO_23?: string;
    CAMPO_24?: string;
    CAMPO_25?: string;
    CAMPO_26?: string;
    CAMPO_27?: string;
    CAMPO_27_1?: string;
    CAMPO_28?: string;
    CAMPO_28_1?: string;
    CAMPO_29?: string;
    CAMPO_29_1?: string;
    CAMPO_30?: string;
    CAMPO_30_1?: string;
    CAMPO_31?: string;
    CAMPO_31_1?: string;
    CAMPO_32?: string;
    CAMPO_32_1?: string;
    CAMPO_33?: string;
    CAMPO_33_1?: string;
    CAMPO_34?: string;
    CAMPO_34_1?: string;
    CAMPO_35?: string;
    CAMPO_36?: string;
    CAMPO_37?: string;
    CAMPO_38?: string;
    CAMPO_39?: string;
    CAMPO_40?: string;
    CAMPO_41?: string;
    CAMPO_42?: string;
    CAMPO_43?: string;
    CAMPO_44?: string;
    CAMPO_45?: string;
    CAMPO_46?: string;
    CAMPO_47?: string;
    CAMPO_48?: string;
    CAMPO_49?: string;
    CAMPO_50?: string;
    CAMPO_51?: string;
    CAMPO_52?: string;
    CAMPO_53?: string;
    CAMPO_54?: string;
    CAMPO_55?: string;
    CAMPO_56?: string;
    CAMPO_57?: string;
    CAMPO_58?: string;
    CAMPO_59?: string;
    CAMPO_60?: string;
    CAMPO_61?: string;
    CAMPO_62?: string;
    CAMPO_62_1?: string;
    CAMPO_62_2?: string;
    CAMPO_62_3?: string;
    CAMPO_62_4?: string;
    CAMPO_62_5?: string;
    CAMPO_62_6?: string;
    CAMPO_62_7?: string;
    CAMPO_62_8?: string;
    CAMPO_62_9?: string;
    CAMPO_62_10?: string;
    CAMPO_62_11?: string;
    CAMPO_63?: string;
    CAMPO_63_1?: string;
    CAMPO_64?: string;
    CAMPO_65?: string;
    CAMPO_66?: string;
    CAMPO_67?: string;
    CAMPO_68?: string;
    CAMPO_69?: string;
    CAMPO_69_1?: string;
    CAMPO_69_2?: string;
    CAMPO_69_3?: string;
    CAMPO_69_4?: string;
    CAMPO_69_5?: string;
    CAMPO_69_6?: string;
    CAMPO_69_7?: string;
    CAMPO_70?: string;
    CAMPO_70_1?: string;
    CAMPO_70_2?: string;
    CAMPO_70_3?: string;
    CAMPO_70_4?: string;
    CAMPO_70_5?: string;
    CAMPO_70_6?: string;
    CAMPO_70_7?: string;
    CAMPO_70_8?: string;
    CAMPO_70_9?: string;
    CAMPO_71?: string;
    CAMPO_72?: string;
    CAMPO_73?: string;
    CAMPO_74?: string;
    CAMPO_75?: string;
    CAMPO_76?: string;
    CAMPO_77?: string;
    CAMPO_78?: string;
    CAMPO_79?: string;
    CAMPO_80?: string;
    CAMPO_80_1?: string;
}